requirements.txt
# Web app
streamlit>=1.36

# PDF / DOCX parsing
pypdf>=3.0
pypdf2
python-docx>=1.1.0
fpdf>=1.7.2

# File conversions & report generation
reportlab>=4.2
pandas>=2.2
openpyxl>=3.1

# LLM / LangChain
langchain-community>=0.1.0
ollama>=0.1.8

# Utilities
requests>=2.32



scikit-learn
streamlit-quill

